% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{.dir_exists}
\alias{.dir_exists}
\title{Check if a give directory exists.}
\usage{
.dir_exists(output_directory)
}
\arguments{
\item{output_directory}{Directory path}
}
\value{
\code{bool} flag indicating if the directory exists.
}
\description{
Check if a give directory exists.
}
