% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{temporal_mosaic}
\alias{temporal_mosaic}
\title{Create mosaic of image based on time.}
\usage{
temporal_mosaic(images_table, output_dir)
}
\arguments{
\item{images_table}{data.frame Table with the following columns:
\itemize{
\item \code{id} (str): Scene ID
\item \code{collection} (str): Scene collection Name
\item \code{date} (Date): Scene date
\item \code{tile} (str): Scene tile
\item \code{file} (str): Valid path to raster file.
}}

\item{output_dir}{str Output directory.}
}
\value{
\code{data.frame} containing the output files
}
\description{
This functions get images reference from a table, group them by date and
create mosaics using the images available in each group.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}
}
